feat: add shared job fetch helper
